mgr/cephadm: move _create_daemon to serve.py
`_create_daemon` can potentially make the CLI unresponsive and should
only be called from the serve() thread.
